All patches ready for 1.19.
v1.9.1
Modified Mandau Pasir to be a fire damage dagger and to be inline with other fire weapons since now daggers have 2 more top tier options.
Increased damage for Sword Whip and Rhava Velar to deal 54 damage with all its hits to be in line with other unique weapons.
For now only Ridil and Oracle Sword deals total 60 damage, will be checking them to see if they need a change or not, next update will be along patch 1.19 for steam.